separate openvino inference script
